requirements.txt
requests
getmac
python-dotenv